Refactor Icon Exports
