+++ packages/constellation/src/components/Base/Icon/Icon.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { SemanticColors } from '../Colors'
 import * as Icons from '.'
-import IconComponent from './Icon'
+import { createIconComponent as IconComponent } from './Icon'
 import type { Props } from './Icon.types'
 
 export const Icon = ({ icon = 'IconPlus', ...args }: Props & { icon: keyof typeof Icons }) => {
